Move MessageBlock and BlockCache types to player/types
These types are used within the PlayerState and should exist in the player types file rather than the random access files so they can be used by players other than random access.
